Updates to how intbgd is created
Used setPars formalism here rather than accessing the parameters directly. For me this sped things up considerably (minutes --> seconds) for the creation of the bgdparams.xcm file.

Also forced all of the line strengths to be relative and tied together rather than independent.
Properly track A/B parameters.

Had to track the parameter offsets wrt to the reference spectrum when mixing FPMA and FPMB.

"Solar" low energy components to be all be free based on presets.

Respects use of fix_line_ratios flag to have the lines be tied. Default I guess is fix_line_ratios = False? (it probably should not be)